requirement.txt
Flask
Pillow
numpy
pytesseract
click
itsdangerous
Jinja2
MarkupSafe
uWSGI
Werkzeug
